Rearrange redis streams
