+++ apps/web/app/(ee)/api/cron/streams/update-workspace-links-usage/route.ts
@@ -2,11 +2,11 @@ import { handleAndReturnErrorResponse } from "@/lib/api/errors";
 import { sendLimitEmail } from "@/lib/cron/send-limit-email";
 import { verifyVercelSignature } from "@/lib/cron/verify-vercel";
 import { WorkspaceProps } from "@/lib/types";
+import { RedisStreamEntry } from "@/lib/upstash/redis-streams/client";
 import {
-  RedisStreamEntry,
   WorkspaceLinksUsageEvent,
   workspaceLinksUsageStream,
-} from "@/lib/upstash/redis-streams";
+} from "@/lib/upstash/redis-streams/workspace-links-usage";
 import { prisma } from "@dub/prisma";
 import { log } from "@dub/utils";
 import { NextResponse } from "next/server";
