+++ apps/web/app/api/links/sync/route.ts
@@ -2,7 +2,7 @@ import { propagateBulkLinkChanges } from "@/lib/api/links/propagate-bulk-link-ch
 import { withWorkspace } from "@/lib/auth";
 import { exceededLimitError } from "@/lib/exceeded-limit-error";
 import { SimpleLinkProps } from "@/lib/types";
-import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams";
+import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams/workspace-links-usage";
 import { prisma } from "@dub/prisma";
 import { NextResponse } from "next/server";
