+++ apps/web/lib/api/links/bulk-create-links.ts
@@ -1,5 +1,5 @@
 import { ProcessedLinkProps } from "@/lib/types";
-import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams";
+import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams/workspace-links-usage";
 import { prisma } from "@dub/prisma";
 import { Prisma } from "@dub/prisma/client";
 import { getParamsFromURL, linkConstructorSimple, truncate } from "@dub/utils";
