+++ apps/web/lib/api/links/create-link.ts
@@ -3,7 +3,7 @@ import { getPartnerEnrollmentInfo } from "@/lib/planetscale/get-partner-enrollme
 import { isNotHostedImage, storage } from "@/lib/storage";
 import { recordLink } from "@/lib/tinybird";
 import { ProcessedLinkProps } from "@/lib/types";
-import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams";
+import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams/workspace-links-usage";
 import { propagateWebhookTriggerChanges } from "@/lib/webhook/update-webhook";
 import { prisma } from "@dub/prisma";
 import { Prisma } from "@dub/prisma/client";
