+++ apps/web/lib/middleware/utils/crawl-bitly.ts
@@ -2,7 +2,7 @@ import { createId } from "@/lib/api/create-id";
 import { linkCache } from "@/lib/api/links/cache";
 import { encodeKeyIfCaseSensitive } from "@/lib/api/links/case-sensitivity";
 import { recordLink } from "@/lib/tinybird";
-import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams";
+import { publishWorkspaceLinksUsageEvent } from "@/lib/upstash/redis-streams/workspace-links-usage";
 import { prisma } from "@dub/prisma";
 import {
   DUB_HEADERS,
