+++ apps/web/lib/upstash/redis-streams/client.ts
@@ -1,4 +1,4 @@
-import { redis } from "./redis";
+import { redis } from "../redis";
 
 export type RedisStreamEntry<T> = {
   id: string;
@@ -137,88 +137,3 @@ export class RedisStream {
     }
   }
 }
-
-/* Workspace Clicks Usage Stream */
-const WORKSPACE_CLICKS_USAGE_UPDATES_STREAM_KEY = "workspace:usage:updates";
-export const workspaceClicksUsageStream = new RedisStream(
-  WORKSPACE_CLICKS_USAGE_UPDATES_STREAM_KEY,
-);
-export interface WorkspaceClicksUsageEvent {
-  linkId: string;
-  workspaceId: string;
-  timestamp: string;
-}
-// Publishes a click event to any relevant streams in a single transaction
-export const publishWorkspaceClicksUsageEvent = async (
-  event: WorkspaceClicksUsageEvent,
-) => {
-  const { linkId, workspaceId, timestamp } = event;
-  try {
-    return await redis.xadd(WORKSPACE_CLICKS_USAGE_UPDATES_STREAM_KEY, "*", {
-      linkId,
-      workspaceId,
-      timestamp,
-    });
-  } catch (error) {
-    console.error("Failed to publish workspace clicks usage event:", error);
-    throw error;
-  }
-};
-
-/* Workspace Links Usage Stream */
-const WORKSPACE_LINKS_USAGE_UPDATES_STREAM_KEY = "workspace:linksUsage:updates";
-export const workspaceLinksUsageStream = new RedisStream(
-  WORKSPACE_LINKS_USAGE_UPDATES_STREAM_KEY,
-);
-
-export interface WorkspaceLinksUsageEvent {
-  workspaceId: string;
-  linksCount: number;
-  timestamp: string;
-}
-
-export const publishWorkspaceLinksUsageEvent = async (
-  event: WorkspaceLinksUsageEvent,
-) => {
-  const { workspaceId, linksCount, timestamp } = event;
-
-  try {
-    return await redis.xadd(WORKSPACE_LINKS_USAGE_UPDATES_STREAM_KEY, "*", {
-      workspaceId,
-      linksCount,
-      timestamp,
-    });
-  } catch (error) {
-    console.error("Failed to publish workspace links usage event:", error);
-    throw error;
-  }
-};
-
-/* Partner Stats Stream */
-const PARTNER_ACTIVITY_STREAM_KEY = "partner:activity:updates";
-export const partnerActivityStream = new RedisStream(
-  PARTNER_ACTIVITY_STREAM_KEY,
-);
-export interface PartnerActivityEvent {
-  programId: string;
-  partnerId: string;
-  timestamp: string;
-  eventType: "click" | "lead" | "sale" | "commission";
-}
-// Publishes a partner activity event to the stream
-export const publishPartnerActivityEvent = async (
-  event: PartnerActivityEvent,
-) => {
-  const { programId, partnerId, timestamp, eventType } = event;
-  try {
-    return await redis.xadd(PARTNER_ACTIVITY_STREAM_KEY, "*", {
-      programId,
-      partnerId,
-      timestamp,
-      eventType,
-    });
-  } catch (error) {
-    console.error("Failed to publish partner activity event to stream:", error);
-    throw error;
-  }
-};
